translations pulled from tx because I want to update AR powerpoint presentation to show everything in spanish. Translation on hometopbar to reflect daily reports. checkWasToday moved to TimeUtils, now importing from correct place in ListOfPatients and deleted its original file
