+++ src/Practitioner/ReviewPatients/HomeTopBar.js
@@ -11,8 +11,7 @@ export default function TopBar() {
   return (
     <Box bgcolor="white" borderBottom="solid 1px lightgray" padding=".5em 1em">
       <Grid alignItems="center" container>
-        {/* Why does it say all patients? */}
-        <PageLabelTitle title={`${t('coordinator.cardTitles.allPatients')}`} />
+        <PageLabelTitle title={`${t('dashboard.dailyReports')}`} />
         <Box flexGrow="1" />
         <IconButton
           style={{
