--- src/Practitioner/ReviewPatients/checkWasToday.js
@@ -1,7 +1,0 @@
-import { DateTime } from 'luxon';
-
-const checkWasToday = (isoTime) => {
-  return DateTime.fromISO(isoTime).toISODate() === DateTime.local().toISODate();
-};
-
-export default checkWasToday;
