+++ src/Utility/TimeUtils.js
@@ -1,14 +1,23 @@
-import {DateTime} from "luxon"
+import { DateTime } from 'luxon';
 
-function daysSinceISODateTime(isoDateTime){
-    return DateTime.fromISO(isoDateTime).diffNow("days").days * -1;
+function daysSinceISODateTime(isoDateTime) {
+  return DateTime.fromISO(isoDateTime).diffNow('days').days * -1;
 }
 
-function daysSincePhotoRequest(isoDateTime){
-    return DateTime.fromISO(isoDateTime).endOf("day").diffNow("days").days * -1;
+function daysSincePhotoRequest(isoDateTime) {
+  return DateTime.fromISO(isoDateTime).endOf('day').diffNow('days').days * -1;
 }
 
-let DATETIME_MED_NO_YEAR = {...DateTime.DATETIME_MED}
-delete DATETIME_MED_NO_YEAR.year
+const checkWasToday = (isoTime) => {
+  return DateTime.fromISO(isoTime).toISODate() === DateTime.local().toISODate();
+};
 
-export {daysSinceISODateTime,daysSincePhotoRequest, DATETIME_MED_NO_YEAR}
+let DATETIME_MED_NO_YEAR = { ...DateTime.DATETIME_MED };
+delete DATETIME_MED_NO_YEAR.year;
+
+export {
+  daysSinceISODateTime,
+  daysSincePhotoRequest,
+  checkWasToday,
+  DATETIME_MED_NO_YEAR,
+};
